* 'develop' of https://github.com/rokwire/illinois-app:
  version: 3.0.70+1070
  Updated CHANGELOG.md (#1146).
  Updated Polls BLE related UI (#1146).
  Android: change the name of the notification channel. [#1146]
  Android: remove enabling bluetotth [#1146]
  Updated NSBluetoothAlwaysUsageDescription and NSBluetoothPeripheralUsageDescription (#1146).
  Remove PollsPlugin and BluetoothServices from Android side (#1146).
  Remove PollsPlugin and BluetoothServices from iOS side (#1146).
  Remove PollsPlugin and BluetoothServices from flutter side (#1146).
  Acknowledged integer latitude/longitude when evaluating explore location distance.

+++ lib/model/Explore.dart
@@ -138,7 +138,7 @@ class ExploreHelper {
       Location location = explore.exploreLocation;
       if (location != null) {
         if ((locationData != null) && (location.latitude != null) && (location.longitude != null)) {
-          double distance = AppLocation.distance(location.latitude, location.longitude, locationData.latitude, locationData.longitude);
+          double distance = AppLocation.distance(location.latitude.toDouble(), location.longitude.toDouble(), locationData.latitude, locationData.longitude);
           return distance.toStringAsFixed(1) + " mi away";
         }
         if ((location.description != null) && location.description.isNotEmpty) {
